pyproject.toml
[project]
name = "agnostudy"
version = "0.1.0"
description = "Learning Agno from the basics with OpenAI-compatible models"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "agno>=1.0.0",
    "beautifulsoup4>=4.12.0",
    "chromadb>=0.5.0",
    "ddgs>=9.13.1",
    "fastapi>=0.136.0",
    "openai>=2.30.0",
    "pydantic>=2.0.0",
    "pypdf>=5.0.0",
    "python-dotenv>=1.0.0",
    "reportlab>=4.0.0",
    "sqlalchemy>=2.0.49",
]
